fix: deep code review pass 18 - 25 issues fixed (#270)
* fix: deep code review pass 18 - 25 issues fixed

Safety & correctness:
- V9 OptionsData::parse() default limit: 10,000 → 1,024 (matches parser pipeline)
- V9 lookup: ~100 IANA-assigned fields (180-322) mapped as IanaAssigned (were Unknown)
- MAX_FIELD_COUNT type: u16 → usize (matches Config field type)
- Config::validate() added for early validation without parser instance
- Config::Default impl added
- Config::with_enterprise_registry() DRY'd via ..Self::new()
- NoTemplateInfo: added Eq impl

API surface cleanup:
- ScopeParser, OptionsFieldParser, FlowSetParser, FieldParser (V9): pub → pub(crate)
- FieldParser (IPFIX): pub → pub(crate)
- DATA_TEMPLATE_V9_ID, OPTIONS_TEMPLATE_V9_ID: pub → pub(crate)
- hook_error_count() delegation added to NetflowParser

CI/build hardening:
- CI: examples now built (cargo build --examples --all-features)
- CI: --no-default-features test added
- CI: benchmarks gated to main/PRs only
- RELEASES.md MSRV corrected (1.87 → 1.88)
- Fuzz crate edition: 2021 → 2024
- Round-trip fuzz target added (fuzz_round_trip)
- Redundant tokio-macros dev-dependency removed

Test coverage:
- V9 options template + data round-trip test
- IPFIX options template + data round-trip test
- Protocol round-trip test (all 256 u8 values)
- Protocol Unknown(145-254) and Reserved(255) test
- test_auto_scoped_parser_template_isolation (V9 templates, not V5)
- test_cache_hit_and_miss_tracking (exact counts, tests misses)
- test_template_ids_after_parsing
- test_hit_rate_after_activity
- test_hooks_fire_during_parsing (via actual parsing path)
- test_hook_error_count

* fix: fuzz

+++ src/variable_versions/ipfix/parser.rs
@@ -39,7 +39,7 @@ impl Default for IPFixParser {
         // Safe to unwrap because DEFAULT_MAX_TEMPLATE_CACHE_SIZE is non-zero
         let config = Config {
             max_template_cache_size: DEFAULT_MAX_TEMPLATE_CACHE_SIZE,
-            max_field_count: usize::from(MAX_FIELD_COUNT),
+            max_field_count: MAX_FIELD_COUNT,
             max_template_total_size: usize::from(u16::MAX),
             max_error_sample_size: 256,
             max_records_per_flowset: DEFAULT_MAX_RECORDS_PER_FLOWSET,
